+++ scripts/ci/pylint_todo.txt
@@ -80,7 +80,7 @@
 ./airflow/contrib/sensors/bash_sensor.py
 ./airflow/contrib/sensors/celery_queue_sensor.py
 ./airflow/contrib/sensors/datadog_sensor.py
-./airflow/contrib/sensors/emr_base_sensor.py
+./airflow/providers/amazon/aws/sensors/emr_base.py
 ./airflow/contrib/sensors/emr_job_flow_sensor.py
 ./airflow/contrib/sensors/emr_step_sensor.py
 ./airflow/sensors/filesystem.py
